+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.55](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.7.54...@times-components/article-label@2.7.55) (2021-05-21)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
 ## [2.7.54](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.7.53...@times-components/article-label@2.7.54) (2021-05-10)
 
 **Note:** Version bump only for package @times-components/article-label