+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.56](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.55...@times-components/error-view@2.3.56) (2021-05-21)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.3.55](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.54...@times-components/error-view@2.3.55) (2021-05-10)
 
 **Note:** Version bump only for package @times-components/error-view