+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.38.18](https://github.com/newsuk/times-components/compare/@times-components/styleguide@3.38.17...@times-components/styleguide@3.38.18) (2021-05-21)
+
+**Note:** Version bump only for package @times-components/styleguide
+
+
+
+
+
 ## [3.38.17](https://github.com/newsuk/times-components/compare/@times-components/styleguide@3.38.16...@times-components/styleguide@3.38.17) (2021-05-10)
 
 **Note:** Version bump only for package @times-components/styleguide