+++ packages/video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.11.6](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.5...@times-components/video@4.11.6) (2021-05-21)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.11.5](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.4...@times-components/video@4.11.5) (2021-05-13)
 
 **Note:** Version bump only for package @times-components/video